Refactor : 공통 사용 코드 분리
